more cleanup
more modularity
moved re-used variables to top of main
everything else is passed in and out
some previous commits may not have been working.  this one does.

and by working, i mean it can encrypt and decrypt, and recognize ping
packets.  however none of this matters as much since the server will
disconnect regardless if the client pings back in a reasonable time,
because its waiting on authentication with the login server -- sending a
Packet93 is not enough.  the guildcard provided is only used to make
sure you are who you say you are and to check if you are banned from the
block.

before any connection to the block is made, you gotta connect to the
login server.  yay.